Add implementation and traversal of a linked list
This commit contains a newly created notebook file explaining the implementation and traversing process of a LinkedList and its optimised version. It covers topics such as creating Node and LinkedList classes, appending nodes, different types of linked lists - singly, doubly & circular, and implementing these Linked lists and their traversal in an efficient way. This notebook can serve as a comprehensive guide for understanding and visualizing the Linked List data structure.